+++ wpa_project/program_app/templates/program_app/class_registration.html
@@ -118,7 +118,7 @@ Wait list is under evaluation and may be discontinued.
     </div>
     <div id="confirm-covid" hidden >
 Covid-19 Policy:
-Please confirm that you agree to bring proof of vaccination proof of a negative PCR Covid-19 test within the previous 48 hours for yourself and all the members of your party.
+Please confirm that you agree to bring proof of vaccination or proof of a negative PCR Covid-19 test within the previous 48 hours for yourself and all the members of your party.
     </div>
 
     {% include "program_app/registration_dialogs.html" %}
