+++ wpa_project/student_app/forms/recaptcha_form.py
@@ -1,5 +1,5 @@
 import requests
-from django.forms import Form, CharField, URLField
+from django.forms import Form, CharField, URLField, HiddenInput
 from django.conf import settings
 
 import logging
@@ -9,10 +9,10 @@ logger = logging.getLogger(__name__)
 class RecaptchaForm(Form):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.fields['captcha'] = CharField()
-        self.fields['captcha'].widget.attrs.update({'class': 'form-control m-2', 'style': 'display:none'})
-        self.fields['url'] = CharField()
-        self.fields['url'].widget.attrs.update({'class': 'form-control m-2', 'style': 'display:none'})
+        self.fields['captcha'] = CharField(widget=HiddenInput)
+        # self.fields['captcha'].widget.attrs.update({'class': 'form-control m-2', 'style': 'display:none'})
+        self.fields['url'] = CharField(widget=HiddenInput)
+        # self.fields['url'].widget.attrs.update({'class': 'form-control m-2', 'style': 'display:none'})
 
     def get_score(self, remote_ip):
         data = {"secret": settings.RECAPTCHA_SECRET_KEY_V3, "response": self.cleaned_data['captcha']}
